customer_service_agent/requirements.txt
langgraph
pymongo
chromadb
fastapi
python-dotenv
langchain-openai
faiss-cpu
langchain-community
uvicorn
pydantic-settings
pypdf
langfuse
tavily-python
